[FE] feat: 로그인 상태에 따라 헤더 정보 수정
- localStorage의 상태에 따라 헤더 다르게 보이도록 수정
- Header를 Header, Nav, Dropdown으로 분리
- Dropdown 추가
