+++ WEB/FE/src/components/LogIn/LogInContainer.tsx
@@ -1,10 +1,11 @@
-import React, { useState, useContext } from 'react';
+import React, { useState } from 'react';
 import { LogInForm } from '@components/LogIn/LogInForm';
 import { TOTPModal } from '@components/TOTPModal/TOTPModal';
 import { loginWithOTP, sendSecretKeyEmail } from '@api/index';
 import { useGoogleReCaptcha } from 'react-google-recaptcha-v3';
 import { useHistory } from 'react-router-dom';
 import { message } from '@utils/message';
+import storageHandler from '@utils/localStorage';
 
 const TOTP_LEN = 6;
 
@@ -29,7 +30,7 @@ const LogInContainer = ({}: LogInContainerProps): JSX.Element => {
     setIsModalOpen(false);
   };
   const onClickModal = () => {
-    setHasTOTPModalError(false); // /secret-key/email
+    setHasTOTPModalError(false);
     executeRecaptcha('sendSecretKeyEmail')
       .then((reCaptchaToken: string) => sendSecretKeyEmail({ authToken, reCaptchaToken }))
       .then(() => alert(message.EMAILSECRETKEYSUCCESS))
@@ -49,14 +50,14 @@ const LogInContainer = ({}: LogInContainerProps): JSX.Element => {
     setModalDisabled(true);
     executeRecaptcha('LogInWithOTP')
       .then((reCaptchaToken: string) => loginWithOTP({ authToken, totp: TOTP, reCaptchaToken }))
-      .then(() => successLoginHandler())
+      .then(({ userName }: { userName: string }) => successLoginHandler(userName))
       .catch((err: any) => onErrorWithOTP(err.response?.data?.message || err.message))
       .finally(() => setModalDisabled(false));
   };
 
-  const successLoginHandler = () => {
+  const successLoginHandler = (userName: string) => {
     alert(message.SIGNINSUCCESS);
-    localStorage.user = 'login';
+    storageHandler.set(userName);
     history.replace('/');
   };
