+++ WEB/FE/src/components/PrivateRoute/PrivateRoute.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Redirect, Route, RouteProps } from 'react-router-dom';
+import storageHandler from '@utils/localStorage';
 
 interface PrivateRouteProps extends RouteProps {}
 
 const PrivateRoute: React.FC<PrivateRouteProps> = (props) => {
-  const isAuthenticated = localStorage.user;
+  const isAuthenticated = storageHandler.get();
 
   return isAuthenticated ? (
     <Route {...props} />
